+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 3.2.0
 
 * Add an option to remove duration from execution summary.
 * Add jasmine & protractor integration tests [#82](https://github.com/bcaudan/jasmine-spec-reporter/issues/82)